Edit deploy

+++ carsello_api/views/event_view.py
@@ -3,7 +3,7 @@ from rest_framework.response import Response
 from rest_framework import serializers, status
 from carsello_api.models import Event
 from carsello_api.permission import AllowSafe
-from rest_framework.decorators import action, permission_classes
+
 
 class EventSerializer(serializers.ModelSerializer):
     class Meta:
@@ -36,6 +36,7 @@ class EventView(ViewSet):
     def create(self, request):
         serialized = CreateEventSerializer(data=request.data)
         serialized.is_valid(raise_exception=True)
+        self.check_object_permissions(request, serialized)
         serialized.save()
         return Response(serialized.data, status=status.HTTP_201_CREATED)
 
@@ -49,10 +50,12 @@ class EventView(ViewSet):
         event.link = request.data['link']
         event.details = request.data['details']
         event.price = request.data['price']
+        self.check_object_permissions(request, event)
         event.save()
         return Response(None, status=status.HTTP_204_NO_CONTENT)
 
     def destroy(self, request, pk=None):
         event = Event.objects.get(pk=pk)
+        self.check_object_permissions(request, event)
         event.delete()
         return Response(None, status=status.HTTP_204_NO_CONTENT)
